Moved award amounts API call to the idv helper
